Feature: Run Uboot in cache
Flow: SPL (EL3), Uboot (EL3), ATF (EL3), Uboot (EL2)

What happens when:
SPL loads Uboot to cache then branches over, Uboot runs in
cache, initializes the main memory and flushes the caches.
Then it branches over to ATF that returns to Uboot switching
the Exception Level from 3 to 2 (EL3 to EL2). Then the system
boots as usual.

In this flavour ATF is linked in to Uboot through the linker script.

Note: this ATF requires a matching Uboot from the Uboot repo from
the branch with the same name.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>