Advisory TFV 8 to CVE-2018-19440.
Use the helper function `save_gp_registers` to save the register
state to cpu_context on entry to EL3 in SMC handler. This has the
effect of saving x0 - x3 as well into the cpu_context which was
not done previously but it unifies the register save sequence
in BL31.

Change-Id: I5753c942263a5f9178deda3dba896e3220f3dd83
Created originally by: Soby Mathew <soby.mathew@arm.com>

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>