Advisory TFV 8 to CVE-2018-19440: "Not saving x0 to x3 registers can leak information from one Normal World SMC client to another"
Use the helper function `save_gp_registers` to save the register
state to cpu_context on entry to EL3 in SMC handler. This has the
effect of saving x0 - x3 as well into the cpu_context which was
not done previously but it unifies the register save sequence
in BL31.

NOTE! Saving all the gp registers is also a mandatory step in preparation
for SMCCC upgrade from v1.0 to v1.1 as the upgraded convention relies on.
This is the reason a compile time option isn't wrapped up around this
advisory.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>