PSCI/power on for secondary cpus: turn MMU on for the secondary cpus.
Do this while keeping the the dcache disabled for the cpus. The cpus
musn't have the dcache enabled on turning the MMU on as the Secure
Monitor runs on the LSM memory which isn't cacheable. Any attempt
at turning MMU on with the dcache on will result in system hang.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>