MEM_PROTECT and RESET2: add PSCI v1.1 generic code APIs
This patch implements the two functionalities MEM_PROTECT and RESET2
defined as additions from PSCI v1.0 to v1.1. MEM_PROTECT provides
protection for cold reboot attacks that exploit the fact that the DDR
memory retains content for several seconds after the removal of power.
The protection is in the form of overriding the whole volatile memory on
cold boot. SYSTEM_RESET2 describes the architectural and vendor-specific
resets. Both of the functions are defined as optional in PSCI v1.1 and
as a results may not have the callbacks registered.

Signed-off-by: Marek Bykowski <marek.bykowski@gmail.com>